Move gql queries to views
