+++ libs/service-portal/assets/src/components/AssetGrid/index.tsx
@@ -1,13 +1,56 @@
 import React, { FC } from 'react'
+import { gql } from '@apollo/client'
 import { useLazyQuery } from '@apollo/client'
 import { useLocale } from '@island.is/localization'
 import { m, TableGrid } from '@island.is/service-portal/core'
 import { Query, UnitsOfUseModel, PropertyLocation } from '@island.is/api/schema'
 import { Text, Box, Button } from '@island.is/island-ui/core'
 import { unitsArray } from '../../utils/createUnits'
-import { GET_UNITS_OF_USE_QUERY } from '../../lib/queries'
 import { DEFAULT_PAGING_ITEMS } from '../../utils/const'
 
+export const GET_UNITS_OF_USE_QUERY = gql`
+  query GetAssetsUnitsOfUse($input: GetPagingTypes!) {
+    assetsUnitsOfUse(input: $input) {
+      paging {
+        page
+        pageSize
+        totalPages
+        offset
+        total
+        hasPreviousPage
+        hasNextPage
+      }
+      unitsOfUse {
+        propertyNumber
+        unitOfUseNumber
+        marking
+        usageDisplay
+        displaySize
+        buildYearDisplay
+        fireAssessment
+        explanation
+        appraisal {
+          activeAppraisal
+          plannedAppraisal
+          activeStructureAppraisal
+          plannedStructureAppraisal
+          activePlotAssessment
+          plannedPlotAssessment
+          activeYear
+          plannedYear
+        }
+        address {
+          locationNumber
+          postNumber
+          municipality
+          propertyNumber
+          display
+          displayShort
+        }
+      }
+    }
+  }
+`
 interface Props {
   units: UnitsOfUseModel
   locationData?: PropertyLocation | null
