+++ libs/service-portal/graphql/src/index.ts
@@ -1,4 +1,5 @@
 export { CompanyInfoFragment } from './lib/fragments/companyInfo'
+export { dataFragment } from './lib/fragments/license'
 export * from './lib/queries/getDocument'
 export * from './lib/queries/listDocuments'
 export * from './lib/queries/getOrganizations'
