--- libs/service-portal/graphql/src/lib/mutations/createEmailVerification.ts
@@ -1,9 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const CREATE_EMAIL_VERIFICATION = gql`
-  mutation createEmailVerification($input: CreateEmailVerificationInput!) {
-    createEmailVerification(input: $input) {
-      created
-    }
-  }
-`
