+++ libs/service-portal/vehicles/src/screens/Overview/Overview.tsx
@@ -24,11 +24,59 @@ import {
 
 import { VehicleCard } from '../../components/VehicleCard'
 import { messages } from '../../lib/messages'
-import { GET_USERS_VEHICLES } from '../../queries/getUsersVehicles'
 import DropdownExport from '../../components/DropdownExport/DropdownExport'
 import { exportVehicleOwnedDocument } from '../../utils/vehicleOwnedMapper'
 import { FeatureFlagClient } from '@island.is/feature-flags'
 import { useFeatureFlagClient } from '@island.is/react/feature-flags'
+
+export const GET_USERS_VEHICLES = gql`
+  query GetUsersVehicles {
+    vehiclesList {
+      persidno
+      name
+      address
+      postStation
+      vehicleList {
+        isCurrent
+        permno
+        regno
+        vin
+        type
+        color
+        firstRegDate
+        modelYear
+        productYear
+        registrationType
+        role
+        operatorStartDate
+        operatorEndDate
+        outOfUse
+        otherOwners
+        termination
+        buyerPersidno
+        ownerPersidno
+        vehicleStatus
+        useGroup
+        vehGroup
+        plateStatus
+        nextInspection {
+          nextInspectionDate
+          nextInspectionDateIfPassedInspectionToday
+        }
+        operatorNumber
+        primaryOperator
+        ownerSsid
+        ownerName
+        lastInspectionResult
+        lastInspectionDate
+        lastInspectionType
+        nextInspectionDate
+      }
+      downloadServiceURL
+      createdTimestamp
+    }
+  }
+`
 
 const defaultFilterValues = {
   searchQuery: '',
