+++ libs/service-portal/vehicles/src/screens/VehicleHistory/VehicleHistory.tsx
@@ -1,6 +1,7 @@
 import isAfter from 'date-fns/isAfter'
 import isEqual from 'lodash/isEqual'
 import React, { useState } from 'react'
+import { gql } from '@apollo/client'
 
 import { useQuery } from '@apollo/client'
 import { Query, VehiclesVehicle } from '@island.is/api/schema'
@@ -24,8 +25,36 @@ import {
 } from '@island.is/service-portal/core'
 
 import { messages } from '../../lib/messages'
-import { GET_USERS_VEHICLES_HISTORY } from '../../queries/getUsersVehicleHistory'
 import TabContent from './TabContent'
+
+export const GET_USERS_VEHICLES_HISTORY = gql`
+  query GetUsersVehicles {
+    vehiclesHistoryList {
+      vehicleList {
+        permno
+        regno
+        type
+        color
+        firstRegDate
+        modelYear
+        productYear
+        role
+        operatorStartDate
+        operatorEndDate
+        outOfUse
+        otherOwners
+        termination
+        vehicleStatus
+        plateStatus
+        nextInspection {
+          nextInspectionDate
+          nextInspectionDateIfPassedInspectionToday
+        }
+        deregistrationDate
+      }
+    }
+  }
+`
 
 const getFilteredVehicles = (
   vehicles: VehiclesVehicle[],
